refactor dpm portfolio memory routes
