+++ src/app/routers/dpm_command_center.py
@@ -7,7 +7,6 @@ from app.contracts.dpm_command_center import (
     DpmExceptionSummaryGatewayResponse,
     DpmExceptionSummaryRequest,
     DpmOutcomeReviewErrorDetail,
-    DpmPortfolioMemoryGatewayResponse,
 )
 from app.middleware.correlation import correlation_id_var
 from app.routers.dpm_openapi import manage_upstream_error_responses
@@ -344,113 +343,3 @@ async def get_mandate_diff(
     )
 
 
-@router.get(
-    "/portfolios/{portfolio_id}/memory",
-    response_model=DpmPortfolioMemoryGatewayResponse,
-    summary="Get DPM portfolio memory",
-    description=(
-        "What: returns the manage-owned RFC-0040/RFC-0041/RFC-0042 portfolio-memory timeline "
-        "for one portfolio, including proof-pack, wave, handoff, and outcome-review lineage. "
-        "When: use this for Workbench event timelines, audit drawers, PM review, and operations "
-        "handoff views that need queryable source-backed portfolio memory. How: Gateway forwards "
-        "the portfolio id and limit to lotus-manage and preserves event order, event types, "
-        "source refs, artifact refs, reason codes, source systems, and content hashes without "
-        "reconstructing timeline nodes or calculating risk, performance, tax, cash, FX, or "
-        "execution truth locally."
-    ),
-)
-async def get_portfolio_memory(
-    portfolio_id: str = Path(
-        ...,
-        description="Core-governed portfolio identifier for the manage-owned memory view.",
-        examples=["PB_SG_GLOBAL_BAL_001"],
-    ),
-    limit: int = Query(
-        default=100,
-        ge=1,
-        le=500,
-        description="Maximum manage-owned memory events to return.",
-        examples=[100],
-    ),
-) -> DpmPortfolioMemoryGatewayResponse:
-    return await dpm_command_center_service().get_portfolio_memory(
-        portfolio_id=portfolio_id,
-        filters={"limit": limit},
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/portfolio-memory/search",
-    response_model=DpmPortfolioMemoryGatewayResponse,
-    summary="Search DPM portfolio memory by persisted source lineage",
-    description=(
-        "What: forwards bounded manage-local portfolio-memory search filters, including source "
-        "system and source type, to lotus-manage. When: use this for source-family posture over "
-        "persisted memory evidence before selecting a portfolio timeline. How: Gateway preserves "
-        "the Manage search payload, applied filters, source counts, reason codes, boundaries, and "
-        "content hashes without querying source-owner stores, discovering the global portfolio "
-        "universe, reconstructing raw source payloads, or claiming OMS, execution, client "
-        "communication, fill, or settlement truth."
-    ),
-)
-async def search_portfolio_memory(
-    portfolio_ids: list[str] | None = Query(
-        default=None,
-        description="Optional repeated portfolio identifiers for bounded persisted memory search.",
-        examples=[["PB_SG_GLOBAL_BAL_001", "PB_SG_GLOBAL_INC_002"]],
-    ),
-    event_type: str | None = Query(
-        default=None,
-        description="Optional manage-owned portfolio-memory event type filter.",
-        examples=["OUTCOME_REVIEW_CREATED"],
-    ),
-    supportability_state: str | None = Query(
-        default=None,
-        description="Optional manage-published supportability state filter.",
-        examples=["READY"],
-    ),
-    source_system: str | None = Query(
-        default=None,
-        description="Optional persisted source-system filter.",
-        examples=["lotus-performance"],
-    ),
-    source_type: str | None = Query(
-        default=None,
-        description="Optional persisted source-type filter.",
-        examples=["PortfolioRealizedTaxSummary:v1"],
-    ),
-    limit: int = Query(
-        default=25,
-        ge=1,
-        le=200,
-        description="Maximum number of persisted memory events to return.",
-        examples=[25],
-    ),
-    offset: int = Query(
-        default=0,
-        ge=0,
-        description="Manage-local offset for bounded persisted memory search.",
-        examples=[0],
-    ),
-    source_scan_limit: int | None = Query(
-        default=None,
-        ge=1,
-        le=1000,
-        description="Optional Manage-local scan cap for source-lineage facet derivation.",
-        examples=[250],
-    ),
-) -> DpmPortfolioMemoryGatewayResponse:
-    return await dpm_command_center_service().search_portfolio_memory(
-        filters={
-            "portfolio_ids": portfolio_ids,
-            "event_type": event_type,
-            "supportability_state": supportability_state,
-            "source_system": source_system,
-            "source_type": source_type,
-            "limit": limit,
-            "offset": offset,
-            "source_scan_limit": source_scan_limit,
-        },
-        correlation_id=correlation_id_var.get(),
-    )
